Fixed a bug where packages greater than Node.js's default package size would cause incomplete commands
To address the bug we have to look for "incomplete" packages that isn't completed with \r\n, and then prepend their value to the next package, and so on, until all commands are complete.

This is mostly an issue for the MOTD (such as on underworld1.no.quakenet.org) and for the welcome replies (001-005).